refact: np.longdouble para os vetores

+++ python/GA/projetofinal/pso/particle.py
@@ -4,9 +4,9 @@ class Particle(object):
 
     def __init__(self,position,velocity,**kwargs):
         assert len(position) == len(velocity), "Posição e velocidade devem ter a mesma dimensões (tamanho)"
-        self.__position = np.array(position)
-        self.__velocity = np.array(velocity)
-        self.pbest_position = np.array(self.position)
+        self.__position = np.array(position,dtype=np.longdouble)
+        self.__velocity = np.array(velocity,dtype=np.longdouble)
+        self.pbest_position = np.array(self.position,dtype=np.longdouble)
 
     def __str__(self):
         return f'{self.__position}'
@@ -25,7 +25,7 @@ class Particle(object):
     def position(self,value):
         assert len(value) == len(self.__position), f'dimensões necessária: {len(self.__position)}'
         if not type(value) is np.ndarray:
-            value = np.array(value)
+            value = np.array(value,dtype=np.longdouble)
         self.__position = value
 
     @property
@@ -36,5 +36,5 @@ class Particle(object):
     def velocity(self,value):
         assert len(value) == len(self.__velocity), f'dimensões necessária: {len(self.__velocity)}'
         if not type(value) is np.ndarray:
-            value = np.array(value)
+            value = np.array(value,dtype=np.longdouble)
         self.__velocity = value
